+++ Public/New-IntuneWin32AppRequirementRule.ps1
@@ -28,10 +28,12 @@ function New-IntuneWin32AppRequirementRule {
         Author:      Nickolaj Andersen
         Contact:     @NickolajA
         Created:     2020-01-27
-        Updated:     2020-01-27
+        Updated:     2021-08-31
 
         Version history:
         1.0.0 - (2020-01-27) Function created
+        1.0.1 - (2021-03-22) Added new minimum supported operating system versions to parameter validation
+        1.0.2 - (2021-08-31) Added new minimum supported operating system versions to parameter validation
     #>    
     [CmdletBinding(SupportsShouldProcess = $true)]
     param(
@@ -42,7 +44,7 @@ function New-IntuneWin32AppRequirementRule {
 
         [parameter(Mandatory = $true, HelpMessage = "Specify the minimum supported operating system version as a requirement for the Win32 app.")]
         [ValidateNotNullOrEmpty()]
-        [ValidateSet("1607", "1703", "1709", "1803", "1809", "1903")]
+        [ValidateSet("1607", "1703", "1709", "1803", "1809", "1903", "1909", "2004", "20H2", "21H1")]
         [string]$MinimumSupportedOperatingSystem,
 
         [parameter(Mandatory = $false, HelpMessage = "Specify the minimum free disk space in MB as a requirement for the Win32 app.")]
@@ -61,46 +63,50 @@ function New-IntuneWin32AppRequirementRule {
         [ValidateNotNullOrEmpty()]
         [int]$MinimumCPUSpeedInMHz
     )
-    # Construct table for supported architectures
-    $ArchitectureTable = @{
-        "x64" = "x64"
-        "x86" = "x86"
-        "All" = "x64,x86"
-    }
-
-    # Construct table for supported operating systems
-    $OperatingSystemTable = @{
-        "1607" = "v10_1607"
-        "1703" = "v10_1703"
-        "1709" = "v10_1709"
-        "1803" = "v10_1803"
-        "1809" = "v10_1809"
-        "1903" = "v10_1903"
-        "1909" = "v10_1909"
-        "2004" = "v10_2004"
-    }
-
-    # Construct ordered hash-table with least amount of required properties for default requirement rule
-    $RequirementRule = [ordered]@{
-        "applicableArchitectures" = $ArchitectureTable[$Architecture]
-        "minimumSupportedOperatingSystem" = @{
-            $OperatingSystemTable[$MinimumSupportedOperatingSystem] = $true
+    Process {
+        # Construct table for supported architectures
+        $ArchitectureTable = @{
+            "x64" = "x64"
+            "x86" = "x86"
+            "All" = "x64,x86"
         }
-    }
 
-    # Add additional requirement rule details if specified on command line
-    if ($PSBoundParameters["MinimumFreeDiskSpaceInMB"]) {
-        $RequirementRule.Add("minimumFreeDiskSpaceInMB", $MinimumFreeDiskSpaceInMB)
-    }
-    if ($PSBoundParameters["MinimumMemoryInMB"]) {
-        $RequirementRule.Add("minimumMemoryInMB", $MinimumMemoryInMB)
-    }
-    if ($PSBoundParameters["MinimumNumberOfProcessors"]) {
-        $RequirementRule.Add("minimumNumberOfProcessors", $MinimumNumberOfProcessors)
-    }
-    if ($PSBoundParameters["MinimumCPUSpeedInMHz"]) {
-        $RequirementRule.Add("minimumCpuSpeedInMHz", $MinimumCPUSpeedInMHz)
-    }
+        # Construct table for supported operating systems
+        $OperatingSystemTable = @{
+            "1607" = "v10_1607"
+            "1703" = "v10_1703"
+            "1709" = "v10_1709"
+            "1803" = "v10_1803"
+            "1809" = "v10_1809"
+            "1903" = "v10_1903"
+            "1909" = "v10_1909"
+            "2004" = "v10_2004"
+            "20H2" = "v10_2H20"
+            "21H1" = "v10_21H1"
+        }
 
-    return $RequirementRule
+        # Construct ordered hash-table with least amount of required properties for default requirement rule
+        $RequirementRule = [ordered]@{
+            "applicableArchitectures" = $ArchitectureTable[$Architecture]
+            "minimumSupportedOperatingSystem" = @{
+                $OperatingSystemTable[$MinimumSupportedOperatingSystem] = $true
+            }
+        }
+
+        # Add additional requirement rule details if specified on command line
+        if ($PSBoundParameters["MinimumFreeDiskSpaceInMB"]) {
+            $RequirementRule.Add("minimumFreeDiskSpaceInMB", $MinimumFreeDiskSpaceInMB)
+        }
+        if ($PSBoundParameters["MinimumMemoryInMB"]) {
+            $RequirementRule.Add("minimumMemoryInMB", $MinimumMemoryInMB)
+        }
+        if ($PSBoundParameters["MinimumNumberOfProcessors"]) {
+            $RequirementRule.Add("minimumNumberOfProcessors", $MinimumNumberOfProcessors)
+        }
+        if ($PSBoundParameters["MinimumCPUSpeedInMHz"]) {
+            $RequirementRule.Add("minimumCpuSpeedInMHz", $MinimumCPUSpeedInMHz)
+        }
+
+        return $RequirementRule
+    }
 }